* origin/main: (265 commits)
  [BUG] fix strip streaming empty-string suffix from DSV4 tool arguments (#29883)
  Fix CuTe DSL DSA paged MQA export (#30627)
  [Refactor] Split DeepSeek-V4 MQALayer into a reusable attention base (#30711)
  [diffusion] refactor: reorganize runtime utility and server_args modules (#30447)
  Fix DSV4 HiSparse SWA tail allocation forwarding (#30408)
  Fix Mistral GSM8K chat eval (#27757)
  [diffusion] doc: update cosmos3 cookbook (#27576)
  [NPU] fix npu import cutlass error (#30454)
  [diffusion] feat: support action output for cosmos3 (#27168)
  Fix immediate profiler step range boundary (#30215)
  [misc] Remove unit test cases that fail the admission criteria (round 3) (#30713)
  [style] Extract init-static values in memory-cache path (#30710)
  [style] Extract init-static values in tokenizer + multimodal path (#30709)
  [style] Extract init-static values in forward path (#30708)
  [style] Extract init-static values in scheduler hot path (#30707)
  [CPU] update fla.cpp to support when num_head_v is not multiples of 16 (#30604)
  [Diffusion] Revert CPU AMX optimizations (#30716)
  Avoid implicit field-based side channel in Scheduler planning (#29408)
  Configurable decode retraction order (#30573)
  Localize cur_batch field in Scheduler to avoid field-based state access (#29407)
  ...

+++ python/sglang/srt/layers/attention/triton_ops/decode_attention.py
@@ -130,7 +130,9 @@ def _fwd_kernel_stage1(
     xai_temperature_len: tl.constexpr,
     PAGE_SIZE: tl.constexpr,
 ):
-    cur_batch = tl.program_id(0)
+    # int64 to avoid overflow of flat offsets into Mid_O when
+    # batch * num_head * max_kv_splits * head_dim exceeds 2**31.
+    cur_batch = tl.program_id(0).to(tl.int64)
     cur_head = tl.program_id(1)
     split_kv_id = tl.program_id(2)
 
@@ -388,7 +390,9 @@ def _fwd_grouped_kernel_stage1(
     USE_PDL: tl.constexpr = False,
     PAGE_SIZE: tl.constexpr = 1,
 ):
-    cur_batch = tl.program_id(0)
+    # int64 to avoid overflow of flat offsets into Mid_O when
+    # batch * num_head * max_kv_splits * head_dim exceeds 2**31.
+    cur_batch = tl.program_id(0).to(tl.int64)
     cur_head_id = tl.program_id(1)
     cur_kv_head = cur_head_id // tl.cdiv(kv_group_num, BLOCK_H)
     split_kv_id = tl.program_id(2)
@@ -684,7 +688,9 @@ def _fwd_kernel_stage2(
     HAS_SINK: tl.constexpr,
     USE_PDL: tl.constexpr = False,
 ):
-    cur_batch = tl.program_id(0)
+    # int64 to avoid overflow of flat offsets into Mid_O when
+    # batch * num_head * max_kv_splits * head_dim exceeds 2**31.
+    cur_batch = tl.program_id(0).to(tl.int64)
     cur_head = tl.program_id(1)
 
     if USE_PDL:
